Code segmentation and ordering - AuthController -
