+++ src/presentation/router/AuthRouter.ts
@@ -1,6 +1,15 @@
 import { Router, Response, NextFunction, Request } from 'express';
 import { AuthenticatedRequest } from '../middleware/AuthMiddleware';
-import { AuthController } from '../controller/AuthController';
+import {
+  getRestaurantOwnerProfile,
+  registerCustomer,
+  registerRestaurantOwner,
+  customerLogin,
+  restaurantOwnerLogin,
+  verifyEmail,
+  uploadProfileImage,
+  updateRestaurantLocation
+} from '../controller';
 import { AuthValidators } from '../validators/AuthValidators';
 import { ValidationMiddleware, SanitizationMiddleware } from '../middleware/index';
 import { DIContainer } from '@/infrastructure/di/DIContainer';
@@ -8,28 +17,32 @@ import multer from 'multer';
 
 export class AuthRouter {
   private router: Router;
-  private upload: multer.Multer;  
+  private upload: multer.Multer;
+  private diContainer: DIContainer;
 
-  constructor(private authController: AuthController) {
+  constructor() {
     const storage = multer.memoryStorage();
-    this.upload = multer({  
+    this.upload = multer({
       storage: storage,
-      limits: { fileSize: 5 * 1024 * 1024 } 
+      limits: { fileSize: 5 * 1024 * 1024 }
     });
-    
+
+    this.diContainer = DIContainer.getInstance();
     this.router = Router();
     this.setupRoutes();
   }
 
-
   private setupRoutes(): void {
-    const authMiddleware = DIContainer.getInstance().authMiddleware;
+    const authMiddleware = this.diContainer.authMiddleware;
 
     this.router.get(
       '/profile/restaurant-owner',
       authMiddleware.authenticate,
       authMiddleware.requireRestaurantOwner,
-      (req: Request, res: Response) => this.authController.getRestaurantOwnerProfile(req as AuthenticatedRequest, res)
+      (req: Request, res: Response) => {
+        const controller = new getRestaurantOwnerProfile(this.diContainer.resolve('getRestaurantOwnerProfileUseCase'));
+        controller.getRestaurantOwnerProfile(req as AuthenticatedRequest, res);
+      }
     );
 
     this.router.post(
@@ -37,7 +50,10 @@ export class AuthRouter {
       SanitizationMiddleware.sanitizeCustomerRegistration(),
       AuthValidators.registerCustomer(),
       ValidationMiddleware.handleValidationErrors(),
-      this.authController.registerCustomer
+      (req: Request, res: Response) => {
+        const controller = new registerCustomer(this.diContainer.resolve('registerCustomerUseCase'));
+        controller.registerCustomer(req, res);
+      }
     );
 
     this.router.post(
@@ -45,12 +61,18 @@ export class AuthRouter {
       SanitizationMiddleware.sanitizeRestaurantOwnerRegistration(),
       AuthValidators.registerRestaurantOwner(),
       ValidationMiddleware.handleValidationErrors(),
-      this.authController.registerRestaurantOwner
+      (req: Request, res: Response) => {
+        const controller = new registerRestaurantOwner(this.diContainer.resolve('registerRestaurantOwnerUseCase'));
+        controller.registerRestaurantOwner(req, res);
+      }
     );
 
     this.router.get(
       '/verify',
-      this.authController.verifyEmail
+      (req: Request, res: Response) => {
+        const controller = new verifyEmail();
+        controller.verifyEmail(req, res);
+      }
     );
 
     this.router.post(
@@ -58,7 +80,10 @@ export class AuthRouter {
       SanitizationMiddleware.sanitizeLoginRequest(),
       AuthValidators.login(),
       ValidationMiddleware.handleValidationErrors(),
-      this.authController.customerLogin
+      (req: Request, res: Response) => {
+        const controller = new customerLogin(this.diContainer.resolve('customerLoginUseCase'));
+        controller.customerLogin(req, res);
+      }
     );
 
     this.router.post(
@@ -66,27 +91,35 @@ export class AuthRouter {
       SanitizationMiddleware.sanitizeLoginRequest(),
       AuthValidators.login(),
       ValidationMiddleware.handleValidationErrors(),
-      this.authController.restaurantOwnerLogin
+      (req: Request, res: Response) => {
+        const controller = new restaurantOwnerLogin(this.diContainer.resolve('restaurantOwnerLoginUseCase'));
+        controller.restaurantOwnerLogin(req, res);
+      }
     );
 
     this.router.post(
       '/profile/image',
-      this.upload.single('profileImage'), 
-      authMiddleware.authenticate,   
+      this.upload.single('profileImage'),
+      authMiddleware.authenticate,
       AuthValidators.validateProfileImage(),
       ValidationMiddleware.handleValidationErrors(),
-      this.authController.uploadProfileImage
+      (req: Request, res: Response) => {
+        const controller = new uploadProfileImage(this.diContainer.resolve('uploadProfileImageUseCase'));
+        controller.uploadProfileImage(req, res);
+      }
     );
+
     this.router.post(
       '/location/restaurant',
       authMiddleware.authenticate,
       SanitizationMiddleware.sanitizeRestaurantLocationUpdate(),
       AuthValidators.updateRestaurantLocation(),
       ValidationMiddleware.handleValidationErrors(),
-      this.authController.updateRestaurantLocation
+      (req: Request, res: Response) => {
+        const controller = new updateRestaurantLocation(this.diContainer.resolve('updateRestaurantLocationUseCase'));
+        controller.updateRestaurantLocation(req, res);
+      }
     );
-
-
   }
 
   public getRouter(): Router {
